requirements.txt
gitpython>=3.1.11
pytz>=2020.1